Refactor Worker into partial monitors
